Adapt to master changes

+++ src/AggregateFunctions/AggregateFunctionMinMax.h
@@ -549,7 +549,7 @@ struct AggregateFunctionsMinMaxDataString
 
     void setImpl(StringRef value, Arena * arena)
     {
-        Int32 value_size = value.size;
+        Int32 value_size = static_cast<Int32>(value.size);
         if (value_size <= MAX_SMALL_STRING_SIZE)
         {
             /// Don't free large_data here.
@@ -563,7 +563,7 @@ struct AggregateFunctionsMinMaxDataString
             if (capacity < value_size)
             {
                 /// Don't free large_data here.
-                capacity = roundUpToPowerOfTwoOrZero(value_size);
+                capacity = static_cast<Int32>(roundUpToPowerOfTwoOrZero(value_size));
                 large_data = arena->alloc(capacity);
             }
